requirements.txt
openai
opencv-python
numpy
Pillow
requests
pytesseract
pydantic